Added exclusion rules for a lot of packages that we don't use
The aws-sdk-for-java keeps growing. All services end up in the same
source tree. With the last versions, even tests are included. We now
see dependencies to aspectj, springworks, junit, etc. All introduced
by services that we don't use. The exclusion rules that were added
by this commit removes a lot of that.